% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCalibrate.R
\name{CCalibrate}
\alias{CCalibrate}
\title{CCalibrate}
\usage{
CCalibrate(
  DSBList,
  Nc = rep(1, length(DSBList)),
  Nx = rep(1000, length(DSBList)),
  Mreads_X = Mreads,
  Mreads_C
)
}
\arguments{
\item{DSBList}{A DSBList formatted input.}

\item{Nc}{A vector describing the number of control cells, or amount of control DNA  added to each sample. The absolute number is not very important.}

\item{Nx}{A vector describing the number of experimental cells, or amount of experimental DNA in each sample. The absolute number is not very important.}

\item{Mreads_X}{A vector of Mreads aligning to experimental genome.}

\item{Mreads_C}{A vector of Mreads aligning to control genome.}
}
\description{
Calibrates a CCseq DSBList.
}
\author{
Will Gittens
}
