% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAverage.R
\name{CCAverage}
\alias{CCAverage}
\title{CCAverage}
\usage{
CCAverage(path.name, exp.name, out.mode = "txt")
}
\arguments{
\item{path.name}{A character string defining folder containing individual FullMaps.}

\item{exp.name}{Manually specify a name for the outputted pool file. Defaults to a concatenation of individual file names.}

\item{out.mode}{output in "txt", "rds", or "both" mode.}
}
\description{
Averages multiple Fullmaps together, by averaging HpM.
}
\author{
Will Gittens
}
