% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCFold.R
\name{CCFold}
\alias{CCFold}
\title{CCFold}
\usage{
CCFold(
  in.mode = "fullmap",
  s.dir = getwd(),
  c.dir,
  exp.name,
  genome = "Cer3H4L2",
  bin.width = 100,
  sc = 80,
  si = 100,
  os = 0,
  mcal = F,
  rDNAcal = F,
  cal.path,
  out.mode1 = "full",
  out.mode2 = "strand",
  plot.mode = T,
  ylims = "auto",
  gene.track = F,
  ARS.track = T,
  names
)
}
\arguments{
\item{in.mode}{Reading in .RDS file or a folder of FullMaps? Choose "rds" or "fullmap".  Defaults to "rds".}

\item{s.dir}{A character string defining the RDS file or folder containing the SAMPLE FullMaps.}

\item{c.dir}{A character string defining the RDS file or folder containing the CONTROL FullMaps.}

\item{exp.name}{Manually specify a name for the experiment.}

\item{genome}{Which genome has the data been aligned to? "Cer3H4L2", "W303" or "hg19". Defaults to "Cer3H4L2".}

\item{bin.width}{What size bins, in bp? Defaults to 100.}

\item{sc}{Parameter which changes the "span" of the smooth. Higher numbers will give a broader smooth.}

\item{si}{Parameter which alters the final number of points of the smooth curve to be plotted. Defaults to 100 bp.}

\item{os}{Do you want to offset the Crick strand relative to the Watson strand prior to binning? Choose 3 for Top2, or 1 for Spo11. Choose 0 for no ofsetting. Defaults to 3.}

\item{out.mode1}{"sparse" or "full" format. Defaults to "full". Defaults to "full"}

\item{out.mode2}{Do you want to sum the Watson and Crick strands ("total") or report them seperately ("strand")? Defaults to "strand".}

\item{plot.mode}{Do you want to plot the binned chromosome maps? (T/F). Defaults to TRUE.}

\item{ylims}{A single number specifying the maximum limit of the y axis, for both Watson and Crick. OR "auto", to automatically define for the entire set.}

\item{gene.track}{Do you want to plot the gene track? This is pretty pointless on whole chromosome plots. Awaiting a more sensible implementation. Defaults to FALSE.}

\item{names}{Can specify your own vector of alias names for the libraries.}
}
\description{
Bins, loess smooths and calculates folds of yeast or human CC-seq data quickly using data.table.
}
\author{
Will Gittens
}
