% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{CCfoldHS.binning}
\alias{CCfoldHS.binning}
\title{CCfoldHS.binning}
\usage{
CCfoldHS.binning(
  files,
  i = 1,
  binwidth.v = c(10, 25, 50, 100),
  work.d = getwd(),
  si = 100,
  plotmode = T,
  ARSline = F,
  ylims = 2,
  control.col = "lightblue",
  mutant.col = "red",
  usesmooth = T
)
}
\arguments{
\item{files}{a vector of filenames for a smoothed hotspot log2(x/y) table}

\item{i}{position in the files vector}

\item{binwidth.v}{A used of the binning binwidths in Kb}

\item{work.d}{working directory}

\item{si}{smooth interval default 100}

\item{plotmode}{Whether to generate a plot}

\item{ARSline}{whether to add ARS positions to the plot}

\item{control.col}{What colour do you want the control}

\item{mutant.col}{What colour do you want the mutant}

\item{usesmooth}{Whether to use smoothed HS data or not. When TRUE plotmode is disabled}
}
\description{
Bins smoothed hotspot log2(x/y) data
}
\examples{
files <- list.files(pattern = "m")
packages.check("doParallel")
cl <- makeCluster(7)
registerDoParallel(cl)
packages.check("GenomicRanges",BiocM=T)
packages.check("ggplot2")
packages.check("ggpubr")
packages.check("GenomicFeatures",BiocM=T)
foreach (i = 1:length(files),.packages=c("nealeLabFunctions","GenomicRanges","ggplot2","ggpubr")) \%dopar\% {
 CCfoldHS.binning(files,i)
 }stopCluster(cl)
}
