% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{CCfoldHS.binpileup}
\alias{CCfoldHS.binpileup}
\title{CCfoldHS.binpileup}
\usage{
CCfoldHS.binpileup(
  p.dir = getwd(),
  binwidth.v = c(100, 50, 25, 10),
  log2FoldMax = 2,
  centromere = T,
  EXT = ".png",
  invert.col = F,
  fixmax = F,
  log2foldoffset = 0,
  colhigh = "red",
  collow = "navy",
  usesmooth = T
)
}
\arguments{
\item{p.dir}{parent directory containing the binData folder}

\item{binwidth.v}{A used of the binning binwidths in Kb}

\item{log2FoldMax}{max and min log2 values to plot}

\item{centromere}{should the output be centered on the centromere}

\item{EXT}{should the output be .png or .pdf}

\item{invert.col}{should the colours be inverted}

\item{fixmax}{should log2fold values higher than the max be fixed to the max. If set to false such values are displayed in gray.}

\item{log2foldoffset}{value to offset log2fold values}

\item{colhigh}{Choose positive colour}

\item{collow}{Choose negative colour}
}
\description{
Pilesup binned smoother log2(x/y) data generated from hotspot tables
}
\examples{
setwd(paste0(getwd(),"/smoothed"));CCfoldHS.binpileup()
}
