% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{CCfoldHS.smoother}
\alias{CCfoldHS.smoother}
\title{CCfoldHS.smoother}
\usage{
CCfoldHS.smoother(input, sc = 80, si = 100, txtout = F)
}
\arguments{
\item{input}{filename for a hotspot log2(x/y) table}

\item{sc}{smoothing constant}

\item{si}{smooth interval default 100}
}
\description{
Smooths a log2(x/y) values from a hotspot column using loess
}
\examples{
packages.check("doParallel")
cl <- makeCluster(8)
registerDoParallel(cl)
foreach (i = 1:length(files),.packages=c("nealeLabFunctions")) \%dopar\% {
CCfoldHS.smoother(files[i])
}stopCluster(cl)
}
