% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCHSplotter.R
\name{CCHSplotter}
\alias{CCHSplotter}
\title{CCHSplotter}
\usage{
CCHSplotter(
  path.name = getwd(),
  names,
  exp.name = "NA",
  plot.mode = "png",
  bg = "white",
  coln = "NormHpMChr",
  smooth = F,
  sc = 10
)
}
\arguments{
\item{path.name}{location of the hotspot tables}

\item{names}{names of the tracks}

\item{plot.mode}{"png" or "pdf"}

\item{bg}{"white" or "none"}

\item{coln}{column to use in the hotspot table}

\item{smooth}{whether to add a loess smooth}

\item{sc}{smooth constant}
}
\description{
Plots Hotspot data from a hotspot table
}
\author{
George Brown
}
