% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCList.R
\name{CCList}
\alias{CCList}
\title{CCList}
\usage{
CCList(path.name)
}
\arguments{
\item{path.name}{A character string defining the folder containing FullMaps.}
}
\description{
Reads a folder of FullMaps into a DSBList, calculates Mreads and creates the DSBListNames.
}
\author{
Will Gittens
}
