% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMap.R
\name{CCMap}
\alias{CCMap}
\title{CCMap}
\usage{
CCMap(
  in.mode = "fullmap",
  path.name = getwd(),
  exp.name,
  genome = "Cer3H4L2",
  phase = "meiosis",
  samples,
  os = 0,
  mcal = F,
  cal.path,
  chrom.,
  pos,
  window.w,
  ylims = 50,
  loci = "manual",
  smooth.mode = "none",
  Fsize = 101,
  win = 50,
  norm.factor = 3,
  gene.track = T,
  ARS.track = T,
  Ty.track = T,
  MNase.track = T,
  scc1.track = T,
  disparity = F,
  out.mode = "png",
  plotdimensions = "wide",
  featureplot = 0,
  names,
  origin.line = F,
  portrait = F,
  basic.mode = F,
  plot.scale = F,
  resolution = 400,
  loci.arrow = F,
  gene.names = "genename"
)
}
\arguments{
\item{in.mode}{Reading in .RDS file or a folder of FullMaps? Choose "rds" or "fullmap".  Defaults to "rds".}

\item{path.name}{A character string defining the RDS file or folder containing FullMaps.}

\item{exp.name}{Manually specify a name for the experiment.}

\item{genome}{Which genome has the data been aligned to? "Cer3H4L2", "W303" or "hg19". Defaults to "Cer3H4L2".}

\item{phase}{"meiosis" or "vegetative"}

\item{samples}{Which samples to plot, as a vector of indexes in the DSBList.}

\item{os}{Do you want to offset the Crick strand relative to the Watson strand prior to binning? Choose 3 for Top2, or 1 for Spo11. Choose 0 for no ofsetting. Defaults to 3.}

\item{chrom.}{Which chromosome to plot?}

\item{pos}{Which coordinate to centre plot on, in bp?}

\item{window.w}{How wide a region to plot in bp?}

\item{ylims}{A single number specifying the maximum limit of the y axis, for both Watson and Crick. OR "auto", to automatically define for each sample.}

\item{loci}{Do you want to manually specify a region ("manual"), a specific single gene ("GENENAME"), or multiple maps centred on a types of features in the AllElementsDUB table (e.g "gene", "CEN", "ARS_consensus_sequence". For more options run CCFeatures() ). You can also provide a vector where the second element is "start", "mid", or "end", to centre the pileup on , for example, the end of a gene.}

\item{smooth.mode}{This option controls whether to apply VariableX ("VX") or Hann window ("HW") smoothing to the broad overlay. Defults to no smoothing.}

\item{Fsize}{Fsize for VariableX (See the documentation for VX function). Only used if smooth.mode == "VX".}

\item{win}{Hanning window width. Only used if smooth.mode == "HW".}

\item{norm.factor}{This can be used to scale the height of the smoothed track (works for Hann  or VariableX smoothing). Defaults to 3.}

\item{gene.track}{Do you want to plot the gene track? Defaults to TRUE.}

\item{out.mode}{"pdf" or "png" file format for output.}

\item{featureplot}{(default=0) changes plot cords to match feature for example c("TOP2","start") plots the at the start position of TOP2}

\item{origin.line}{Do you want to plot a dotted line in the centre? Useful for demonstrating offset at finescale resolution.}

\item{plot.scale}{Do you want to keep the overall plot dimensions constant? If so set to "fixed". Defaults to unfixed, which means that the overal height of the plot increases with more samples. This was George's solution to the overlapping label issue.}

\item{loci.arrow}{Do you want to plot an arrow highlighting the centre of the plot? E.g the locus}

\item{gene.names}{Do you want the gene track to be labelled with "genename" (e.g TEL1), or "sysname" (e.g YBL088C)?}

\item{plot.dimensions}{Specify the plot size. Can be any size from the "A sizing" scale (e.g "A1"). However this is not implemented properly yet. Defaults to "wide", which is fine.}
}
\description{
Maps CCs in a specific chromosome region, alongside various annotations.
}
\author{
Will Gittens,George Brown
}
