% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCMask.R
\name{CCMask}
\alias{CCMask}
\title{CCMask}
\usage{
CCMask(path.name = getwd(), in.mode = "fullmap")
}
\arguments{
\item{path.name}{A character string defining the RDS file or folder containing FullMaps.}

\item{in.mode}{Reading in .RDS file or a folder of FullMaps? Choose "rds" or "fullmap".  Defaults to "rds".}
}
\description{
removes mtDNA,rDNAand 2-micron plasmid from Fullmaps
}
\author{
Will Gittens
}
