% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCOffsetfilter.R
\name{CCOffset.filter}
\alias{CCOffset.filter}
\title{CCOffset.filter}
\usage{
CCOffset.filter(path.name = getwd(), offset = 3)
}
\arguments{
\item{path.name}{path to fullmaps}

\item{offset}{choose the range of offsets to analyse (3 for Top2, 1 for Spo11)}
}
\description{
Filters Fullmaps by an offset value
}
\examples{
CCOffset()
}
\author{
Will Gittens, George Brown
}
