% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCOffset.R
\name{CCOffset}
\alias{CCOffset}
\title{CCOffset}
\usage{
CCOffset(
  path.name = getwd(),
  exp.name = "NA",
  os.range = seq(-10, 10, by = 1),
  mode = "WC"
)
}
\arguments{
\item{path.name}{path to fullmaps}

\item{exp.name}{name of experiment}

\item{os.range}{range to plot the offset over}

\item{mode}{specify whether to analyse offsets of Crick relative to Watson ("WC"), or Watson relative to Watson ("WW")}
}
\description{
Generates plots of a pearson offset analysis plots from CCseq fullmaps
}
\examples{
CCOffset()
}
\author{
Will Gittens, George Brown
}
