% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCPileup.R
\name{CCPileup}
\alias{CCPileup}
\title{CCPileup}
\usage{
CCPileup(
  path.name,
  exp.name,
  in.mode = "fullmap",
  loci = "TSS",
  genome = "Cer3H4L2",
  phase = "meiosis",
  samples,
  window.w = 5000,
  combine = F,
  strat.mode = "none",
  strat.levels = 4,
  out.mode = "png",
  ylims = "auto",
  win = 50,
  plot.dimensions = "wide",
  mirror.mode = F,
  strand.sep = F,
  plot.w,
  positive.strands = T,
  normalize = T,
  cache.search = T,
  names,
  col,
  rDNA = F,
  mtDNA = F
)
}
\arguments{
\item{path.name}{A character string defining the RDS file or folder containing FullMaps.}

\item{exp.name}{Manually specify a name for the experiment.}

\item{in.mode}{Reading in .RDS file or a folder of FullMaps? Choose "rds" or "fullmap".  Defaults to "rds".}

\item{loci}{The path to an RDS file or .txt file containing a table of loci in column format: c("Chr", "Pos", "Strand"). Alternatively can use inbuilt loci such as "TSS" (From AllElementsDUB), or "CTSS" (From YeastTSS), "CEN" (Centromeres). More info on the generation of these sets is available from Will.}

\item{genome}{Which genome has the data been aligned to? "Cer3H4L2", "W303" or "hg19". Defaults to "Cer3H4L2".}

\item{phase}{Which phase is it? "meiosis" or "vegetative"}

\item{samples}{Which samples to plot, as a vector of indexes in the DSBList.}

\item{window.w}{How wide a region to pileup in bp?}

\item{combine}{Do you want to plot all samples on the same plot? If so, it is recommended to set strat.mode to "none", or the plots will be very busy. Defaults to FALSE}

\item{strat.mode}{How do you want to stratify the plots? "none" will leave the Pileups unstratified. "manual" will point the stratifier to a 4th numeric column (column name must be "Value") in the user-defined loci table. "expression" or "length" can be used in combination with the inbuilt TSS or CTSS loci tables, and refer to microarray gene expression, and gene length, respectively.}

\item{strat.levels}{How many strata do you want to use? Defaults to 4.}

\item{out.mode}{"pdf" or "png" file format for output.}

\item{ylims}{A single number specifying the maximum limit of the y axis, for both Watson and Crick. OR "auto", to automatically define for each sample.}

\item{win}{Hanning window width for smoothing.}

\item{plot.dimensions}{Specify the plot size. Can be any size from the "A sizing" scale (e.g "A1"). However this is not implemented properly yet. Defaults to "wide", which is fine.}

\item{mirror.mode}{Do you want to average the plot either side of the loci? This generates a symmetrical plot, which is useful when considering loci with no inherent directionality (e.g centromeres). Defaults to FALSE.}

\item{strand.sep}{Do you want to plot the Top and Bottom strands (relative to locus) separately? Defaults to FALSE}

\item{plot.w}{Specify the plot range in bp. Defaults to the pileup window.w.}

\item{normalize}{Do you want to normalize every pileup over the range of window.w? (NOTE: IT NORMALISES OVER window.w , not plot.w!)}

\item{cache.search}{Do you want to search in the folders surrounding the path.name for a matching saved PileUp? Defaults to true.}

\item{positive.strand}{Do you want to plot both Top and Bottom strands on the same positive axis? If so, TRUE. If you want to plot Top on positive and bottom on negative, then change to FALSE.}
}
\description{
Maps CCs in a specific chromosome region, alongside various annotations.
}
\author{
Will Gittens
}
