% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCPipeline.R
\name{CCPipeline}
\alias{CCPipeline}
\title{CCPipeline}
\usage{
CCPipeline(
  work.d,
  exp.name,
  combine = T,
  readMidpoint = F,
  subsamplereads = F,
  R2Map = F
)
}
\arguments{
\item{work.d}{location of the desired working directory.}

\item{exp.name}{name of the rds file generated.}

\item{readMidpoint}{Do you want to call the midpoint of the molecule}

\item{subsamplereads}{Do you want to subsample. For example enter 0.5 to sub sample by 50 percent}

\item{R2Map}{Do you want to generate a R2 map}
}
\description{
Inputs the "Molecule" files in generated in the "Analysis" folder by termMapper.
Outputs whole genome "FullMap" files for each sample, with and without de-duplication (based
identical R1&R2 coordinates between independent molecules).
Also outputs histograms plots which detail the degree of PCR duplication in the each sample.
Also outputs a "DSBList" (list of all FUllMaps), in RDS format.
}
\author{
Will Gittens George Brown
}
