% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSearch.R
\name{CCSearch}
\alias{CCSearch}
\title{CCSearch}
\usage{
CCSearch(path.name, file.name)
}
\arguments{
\item{path.name}{Path to the folder around which you wish to search for the file.}

\item{file.name}{The file name to search for.}
}
\description{
Looks locally around path.name for a file matching file.name, and returns it's path.
}
\author{
Will Gittens
}
