% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSeqbias2.R
\name{CCSeqbias2}
\alias{CCSeqbias2}
\title{CCSeqbias2}
\usage{
CCSeqbias2(
  path.name = getwd(),
  minThresholdv = c(0, 1, 5),
  maxThreshold = "MAX",
  w = 20,
  rDNA = F,
  chroms = c(1:16),
  ylims = c(0.05, 0.6),
  dyad = 1.5,
  genome = "Cer3H4L2",
  plotT = "o"
)
}
\arguments{
\item{path.name}{point to a folder containing fullmaps}

\item{minThresholdv}{A vector containing the min threshold}

\item{maxThreshold}{A value containing the max}

\item{w}{Window width in bp}

\item{rDNA}{include rDNA}

\item{chroms}{A Vector of chromosomes to process.}

\item{ylims}{min max of ylims. E.g c(0.05,0.6)}

\item{dyad}{Dyad to plot center lines from (0.5 for Spo11, 1.5 for Top2)}

\item{genome}{location to genome.fa}

\item{plotT}{#PlotType: "l", "p", "o"}

\item{loci}{this can be set to a specific genename. The sequence bias will be calcuated only for sites within the genebody of this gene. HpM thresholding occurs before this, but you will probably still need to modify min max thresholds.}
}
\description{
generates a sequence bias plot. This version is Will's work in progress, to make a faster, and more generic script.
}
\author{
Will Gittens, George Brown, Matt Neale
}
