% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSplit.R
\name{CCSplit}
\alias{CCSplit}
\title{CCSplit}
\usage{
CCSplit(path.name = getwd(), in.mode = "fullmap")
}
\arguments{
\item{path.name}{A character string defining the RDS file or folder containing FullMaps.}

\item{in.mode}{Reading in .RDS file or a folder of FullMaps? Choose "rds" or "fullmap".  Defaults to "rds".}
}
\description{
Splits fullmaps into seprate chromosome-seperated maps
}
\author{
Will Gittens
}
