% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctools-package.R
\docType{package}
\name{cctools-package}
\alias{cctools-package}
\alias{cctools}
\title{Tools for the continuous convolution trick in nonparametric estimation}
\description{
Implements the uniform scaled beta distribution \code{\link[=dusb]{dusb()}}, a generic function
for continuous convolution \code{\link[=cont_conv]{cont_conv()}}, and the continuous convolution
kernel density estimator \code{\link[=cckde]{cckde()}}.
}
\references{
Nagler, T. (2017). Nonparametric estimation of probability
densities when some variables are discrete. Unpublished manuscript.
}
\author{
Thomas Nagler
}
\keyword{package}
