% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCWrite.R
\name{CCWrite}
\alias{CCWrite}
\title{CCWrite}
\usage{
CCWrite(DSBList. = DSBList, filenames = DSBListNames)
}
\arguments{
\item{DSBList.}{A DSBList object.}

\item{filenames}{The names of the files. Defaults to DSBListNames.}
}
\description{
Write a DSBList into separate FullMap fies.
}
\author{
Will Gittens
}
