% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsb.R
\name{dupsb}
\alias{dupsb}
\alias{rupsb}
\title{UPSB distribution}
\usage{
dupsb(x, b = 0, ell = 5)

rupsb(n, b = 0, ell = 5)
}
\arguments{
\item{x}{vector of quantiles.}

\item{b}{scale parameter of the UPSB distribution.}

\item{ell}{smoothness parameter of the UPSB distribution.}

\item{n}{number of observations.}
}
\description{
The uniform plus scaled beta (UPSB) distribution describes the distribution of
the random variable
\deqn{U_{b, \ell} = U + b(B - 0.5),}
where \eqn{U} is a \eqn{U[-0.5, 0.5]} random variable, \eqn{B} is a
\eqn{Beta(\ell, \ell)} random variable, and \eqn{b > 0, \ell >= 1}.
}
\examples{
# plot distribution
sq <- seq(-0.8, 0.8, by = 0.01)
plot(sq, dupsb(sq), type = "l")
lines(sq, dupsb(sq, b = 0.25), col = 2)
lines(sq, dupsb(sq, b = 0.25, ell = 10), col = 3)

# simulate from the distribution
x <- rupsb(100, b = 0.3, ell = 0)

}
\references{
Nagler, T. (2017).
Nonparametric estimation of probability densities when some variables are
discrete.
Unpublished manuscript.
}
