% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{expand_names}
\alias{expand_names}
\title{Expands names for expand_as_numeric}
\usage{
expand_names(x)
}
\arguments{
\item{x}{as in \code{\link[=expand_as_numeric]{expand_as_numeric()}}.}
}
\value{
A vector of size \code{ncol(expand_as_numeric(x))}.
}
\description{
Expands each element according to the factor expansions of columns in
\code{\link[=expand_as_numeric]{expand_as_numeric()}}.
}
