% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{expand_vec}
\alias{expand_vec}
\title{Expand a vector like expand_as_numeric}
\usage{
expand_vec(y, x)
}
\arguments{
\item{y}{a vector of length 1 or \code{ncol(x)}.}

\item{x}{as in \code{\link[=expand_as_numeric]{expand_as_numeric()}}.}
}
\value{
A vector of size \code{ncol(expand_as_numeric(x))}.
}
\description{
Expands each element according to the factor expansions of columns in
\code{\link[=expand_as_numeric]{expand_as_numeric()}}.
}
