% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureplotter.R
\name{featureplotter}
\alias{featureplotter}
\title{featureplotter}
\usage{
featureplotter(
  x,
  coln = "genename",
  col = "wheat",
  lty = 1,
  strand = "+",
  av = 50,
  lab = T,
  tv = av,
  pos1 = xl1,
  pos2 = xl2,
  cex.size = 0.9
)
}
\arguments{
\item{x}{Featurelist subsetted from the AllElementsDUB}

\item{col}{The fill colour for the feature arrows.}

\item{lty}{Line type for border. 1 for solid, 2 for dotted.}

\item{strand}{Which strand are the features on? can be "+", "-" or "*"}

\item{av}{Vertical location at which to plot feature.}

\item{tv}{Vertical location at which to plot text. Default is same as av (plots within feature).}
}
\description{
Wrapper for Matt's gene track plotter in the original mapper.
}
\author{
Matt Neale and Will Gittens
}
