% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccOutliers.R
\name{findOutlier}
\alias{findOutlier}
\title{Identify outliers using the ROUT method
REF: https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-7-123}
\usage{
findOutlier(dat, Q = 0.05, outliers = F)
}
\arguments{
\item{dat}{The data object as loaded by loadData()}

\item{Q}{FDR trheshold - default 0.05}

\item{outliers}{Boolean indicating if outliers should be kept and returned (T) or removed (F). Default is FALSE.}
}
\value{
A data frame with outliers for each genotype/conditoin pairing removed (outliers=F).
}
\description{
Identify outliers using the ROUT method
REF: https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-7-123
}
