% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullMap2bedGraph.R
\name{FullMap2bedGraph}
\alias{FullMap2bedGraph}
\title{FullMap2BedGraph}
\usage{
FullMap2bedGraph(
  path.name,
  scaling.factor = 1,
  out.dir = getwd(),
  chrom = 1:18,
  mode = "base"
)
}
\arguments{
\item{path.name}{Character string defining the pathname to the fullmap file. Files must be in the format SAMPLENAME.FullMap.txt}

\item{scaling.factor}{a scaling factor used to boost the values. This is useful in calibration, for example. Defaults to 1.}

\item{out.dir}{A character string defining the pathname to the output directory. Defaults to current working directory.}

\item{chrom}{A vector defining the chromosomes to process. Defaults to all chromosomes, including chromosome 3, 17 and 18, which may cause compatibility issues with UCSC.}

\item{mode}{Can be "break" or "base", depending on whether you want to output the position of the DNA break (between bases), or the position of the base which is covalently linked to the protein.}
}
\description{
Converts FullMaps downloaded from GEO (SPECIFICALLY), into bedGraphs which can be uploaded as custom tracks to UCSC.
}
\author{
Will Gittens
}
