% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hannSlide.R
\name{hannSlide}
\alias{hannSlide}
\title{hannSlide}
\usage{
hannSlide(vec, win = 50)
}
\arguments{
\item{vec}{Input vector or (or data table column) to smooth.}

\item{win}{Width of hanning window in bp.}
}
\description{
Smooth an input vector (or data table column), with a sliding Hann window of defined width.
}
\author{
Will Gittens
}
