% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{HS.totals}
\alias{HS.totals}
\title{HS.totals}
\usage{
HS.totals(
  work.d = getwd(),
  outdir = getwd(),
  Pan,
  use.BG = T,
  readfilename = "FullMap.",
  mode = "txt",
  exp.name = "NA",
  use.mreads = T,
  geneTable =
    "/Users/georgebrown/Dropbox/Annotations/AllElementsDUB_Cer3H4L2_Brar_2016.08.16.txt",
  neeman = F,
  WCequal = F,
  extend = 300
)
}
\arguments{
\item{outdir}{output directory}

\item{use.BG}{whether to use BGreads or not}

\item{readfilename}{name of the read pattern}

\item{exp.name}{experiment name}

\item{geneTable}{location of the AllElementsDUB file}

\item{word.d}{Working directory where the fullmaps are.}
}
\description{
Generates a hotspot table from a full map file or rds
set the working directory to the location of the full map files
}
\examples{
HS.totals(work.d=getwd(),outdir="/Users/georgebrown/Desktop/foldhsdev"
,Pan="/Users/georgebrown/Dropbox/Annotations/Cer3H4L2/Pan.Hotspots.IGR.SacCer3_H4L2_2016.08.10a.txt")
}
