% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccLoadData.R
\name{loadData}
\alias{loadData}
\title{Load data table}
\usage{
loadData(datname = NA, refgenotype = NA, naivelevel = "N")
}
\arguments{
\item{datname}{File name of tab delimited output of analearn. Leave NA to choose file in file browser.}

\item{refgenotype}{Character indicating the reference genotype (column name: 'genotype'). Leave NA to select interactively.}

\item{naivelevel}{Character indicating the naive condition (column name: 'condition'). Default is 'N'.}
}
\value{
A data table with genotype and conditions as factors releveled to the indicated reference conditions.
}
\description{
A function that loads in your analearn output file and sets the reference genotypes and conditions
}
