% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAnnotate.R
\name{loadRData}
\alias{loadRData}
\title{loadRData}
\usage{
loadRData(fileName)
}
\arguments{
\item{...}{a RData or RDS file and returns it. This is useful for reassigning the name iupon loadin}
}
\description{
loads a RData file attached in th4e data folder of the package and returns it. This is useful for reassigning the name iupon loadin.
}
