% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRDataG.R
\name{loadRDataG}
\alias{loadRDataG}
\title{loadRDataG}
\usage{
loadRDataG(fileName)
}
\arguments{
\item{loads}{a RData or RDS file and returns it. This is useful for reassigning the name iupon loadin.
hey}
}
\description{
loads a RData or RDS file and returns it. This is useful for reassigning the name iupon loadin.
}
