% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MreadR.R
\name{MreadR}
\alias{MreadR}
\title{MreadR}
\usage{
MreadR(x)
}
\arguments{
\item{x}{A DSBList object, or a path to a DSBList saved in RDS format, expressed as a character string.}
}
\description{
Calculates the total number of reads in a DSBList, outputted as a vector of length equal to the length of the DSBList.
}
\author{
Will Gittens 2020
}
