% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon2.R
\name{polygon2}
\alias{polygon2}
\title{polygon2}
\usage{
polygon2(x, y, ...)
}
\arguments{
\item{x}{x axis coordinate}

\item{y}{y axis coordinate}
}
\description{
wrapper for polygon which makes it behave like other types of base R plotting
}
\author{
Will Gittens
}
