% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounddown.R
\name{rounddown}
\alias{rounddown}
\title{rounddown}
\usage{
rounddown(x, base)
}
\arguments{
\item{x}{Input number.}

\item{base}{The base number}
}
\description{
Rounds down a number to the specified base.
}
\author{
Will Gittens
}
