% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundup.R
\name{roundup}
\alias{roundup}
\title{roundup}
\usage{
roundup(x, base)
}
\arguments{
\item{x}{Input number.}

\item{base}{The base number}
}
\description{
Rounds up a number to the specified base.
}
\author{
Will Gittens
}
