% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{smooth_plotter}
\alias{smooth_plotter}
\title{smooth_plotter}
\usage{
smooth_plotter(work.d = getwd())
}
\arguments{
\item{work.d}{path to smoothed.Rbin files}
}
\description{
Plots all the log2fold smoothed datasets on the same plot
}
