% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs.R
\name{strainTimeSplit}
\alias{strainTimeSplit}
\title{strainTimeSplit}
\usage{
strainTimeSplit(input, timeString = "hr", toADD = "_")
}
\arguments{
\item{input}{a string}

\item{toADD}{character to add}
}
\description{
Adds a character in a string between the time and strain
}
