% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSeqbias.R
\name{vec2str}
\alias{vec2str}
\title{chrom.vector.rename}
\usage{
vec2str(input)
}
\description{
renames a vector of numbers to a short hand string
}
\examples{
chrom.vector.rename(c(1:16))
}
\author{
George Brown
}
