% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VX.R
\name{VX}
\alias{VX}
\title{VX}
\usage{
VX(x, Fsize = 101, norm.factor = 3)
}
\arguments{
\item{x}{A sparse CCseq dataframe or datatable (colnames = c("Chr", "Pos", "Watson", "Crick)). E.g the "sae2.0" table generated within mapper.}

\item{Fsize}{The number of single strand sites (non-zero "Watson" or "Crick" rows) or double-stranded sites ("Total" rows) over which to smooth. This is the number of values that will be averaged, NOT the  distance in bp.}
}
\description{
VariableX smoother which takes a sparse CCseq dataframe or datatable (colnames = c("Chr", "Pos", "Watson", "Crick)), and smooths both the y dimension ("Watson" and "Crick"), and the x dimension ("Pos"). A sliding mean height and mean position of Fsize number of hits is calculated.
}
\author{
Matt Neale and Will Gittens
}
