% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.it.R
\name{wrap.it}
\alias{wrap.it}
\title{wrap.it}
\usage{
wrap.it(x, n)
}
\arguments{
\item{x}{A character string (e.g an axis label)}

\item{n}{The number of characters on each line.}
}
\description{
wraps a character string (x) every n characters
}
\author{
Will Gittens
}
