% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccBoxplots.R
\name{ccBoxplots}
\alias{ccBoxplots}
\title{Courtship boxplots
A function to plot courtship conditioning results as boxplots with pairwise Wilcoxon Rank Sum tests}
\usage{
ccBoxplots(dat)
}
\arguments{
\item{dat}{A data frame generated from loadData() followed by findOutlier()}
}
\value{
A boxplot with pairiwise Wilcoxon Rank Sum test p-values
}
\description{
Courtship boxplots
A function to plot courtship conditioning results as boxplots with pairwise Wilcoxon Rank Sum tests
}
