% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctools-package.R
\docType{package}
\name{cctools-package}
\alias{cctools-package}
\alias{cctools}
\title{Tools for the continuous convolution trick in nonparametric estimation}
\description{
Implements the uniform scaled beta distribution \code{\link[=dusb]{dusb()}}, a generic function
for continuous convolution \code{\link[=cont_conv]{cont_conv()}}, and the continuous convolution
kernel density estimator \code{\link[=cckde]{cckde()}}.
}
\references{
Nagler, T. (2017). \emph{A generic approach to nonparametric function
estimation with mixed data.} \href{https://arxiv.org/pdf/1704.07457.pdf}{arXiv:1704.07457}
}
\author{
Thomas Nagler
}
\keyword{package}
