% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_conv.R
\name{cont_conv}
\alias{cont_conv}
\title{Continuous convolution}
\usage{
cont_conv(x, b = 0, ell = 5)
}
\arguments{
\item{x}{data; numeric matrix or data frame.}

\item{b}{scale parameter of the UPSB distribution (see, \code{\link[=dupsb]{dupsb()}}).}

\item{ell}{smoothness parameter of the UPSB distribution (see, \code{\link[=dupsb]{dupsb()}}).
The estimator uses the Epanechnikov kernel for smoothing and the UPSB for
continuous convolution (default parameters correspond to the
\eqn{U[-0.5, 0.5]} distribution).}
}
\value{
Continuosly convoluted random vector.

The UPSB distribution (\code{\link[=dupsb]{dupsb()}}) is used as the noise distribution. Discrete
variables are assumed to be integer-valued.
}
\description{
Applies the continuous convolution trick, i.e. adding continuous noise to
all discrete variables. If a variable should be treated as discrete, they must
be declared as \code{\link[=ordered]{ordered()}}.
}
\examples{
Z <- rbinom(100, 6, 0.3)  # discrete variable
X <- rexp(1000, 5)        # continuous variable
dat <- cbind(Z, X)

# plot original and continuously convoluted data
plot(dat)
points(cont_conv(dat), col = 2)

}
\references{
Nagler, T. (2017).
Nonparametric estimation of probability densities when some variables are
discrete.
Unpublished manuscript.
}
